% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.date.R
\name{choose.date}
\alias{choose.date}
\title{Choose the time period of wastewater data.}
\usage{
choose.date(data, start_date, end_date)
}
\arguments{
\item{data}{A wastewater dataset with zipcode information.}

\item{start_date}{A string, starting time.}

\item{end_date}{A string, end time.}
}
\value{
Wastewater data within specific time range.
}
\description{
This function is used for setting the time range of wastewater data.
}
\examples{
library(wastewater.analyst)
a="2022-01-01"
b="2022-03-06"
data=read.csv(url('https://raw.githubusercontent.com/jeffbrennan/COVID-19/master/tableau/wastewater_zip.csv'))
df<-choose.date(data,a,b)
#output the data
df
}
