% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cohzip}
\alias{cohzip}
\title{The ZipCode Of City Of Houston}
\format{
A vector with 105 integers:
}
\usage{
cohzip
}
\description{
A dataset containing the 105 zip codes of city of houston
}
\keyword{datasets}
