% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.cal.R
\name{pop.cal}
\alias{pop.cal}
\title{Calculate the weighted sum of Spline_WW_weight in wastewater zip dataset. Also, it will show you a line plot of it.}
\usage{
pop.cal(data, datebreaks = "1 month")
}
\arguments{
\item{data}{A wastewater dataset, with variable date, population, zipcodes, Spline_WW_weight, ZIPCODE}

\item{datebreaks}{The breaks of date you want to show on the picture. The default is 1 month. You should type in it directly like \emph{"two month"}}
}
\value{
Wastewater data with additional variable: Spline_WW_weight_final, which represents the weighted sum weight. Also return the population of each
date.
}
\description{
This function is used for calculating the weighted sum of Spline_WW_weight variable.
The weighted is defined by area_population/total_population
}
\details{
The dataset must be the dataset from \url{'https://raw.githubusercontent.com/jeffbrennan/COVID-19/master/tableau/wastewater_zip.csv'}.
If want to use your own dataset, it must contains the variabels of following names: date, pop, Spline_WW_weight,ZIPCODE.
}
\examples{
library(wastewater.analyst)
# Infile the most up to date version of wastewater zip data
df=update()
#work out the data we need
final_waste<-pop.cal(df)
#output the data
final_waste
}
