% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.zip.R
\name{split.zip}
\alias{split.zip}
\title{Split the wastewater data according to zipcodes.}
\usage{
\method{split}{zip}(data)
}
\arguments{
\item{data}{A wastewater dataset, with variable date, population, zipcodes, Spline_WW_weight}
}
\value{
A list of data frames. The data frames is the wastewater information of each zip code area, ordered by date.
}
\description{
This function will split the wastewater data into different parts according to the zipcodes.
}
\details{
The dataset must be the dataset from \url{'https://raw.githubusercontent.com/jeffbrennan/COVID-19/master/tableau/wastewater_zip.csv'}.
If want to use your own dataset, it must contains the variabels of following names: date, pop, Spline_WW_weight, ZIPCODE.
}
\examples{
library(wastewater.analyst)
# Read the most up to date version of wastewater zip data
df=update()
#work out the data we need
zip_group=split.zip(df)
#output the data in area '77002'
zip_group$`77002`
#See what zipcodes you have in your result
names(zip_group)
}
