% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Wastewater data reading.}
\usage{
update()
}
\value{
Wastewater dataset with ZipCode information.
}
\description{
This function is used for downloading the data from website \url{https://raw.githubusercontent.com/jeffbrennan/COVID-19/master/tableau/wastewater_zip.csv} created by \strong{Jeff Brennan}.
}
\details{
Just use this function \strong{direcly}. It will download the waste water data from the website.
}
\examples{
library(wastewater.analyst)
df<-update()
#output the data
df
}
